% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{SEM}
\alias{SEM}
\title{Standard Error of the Mean}
\usage{
SEM(x)
}
\arguments{
\item{x}{the data set for which the summary statistics is desired}
}
\description{
This function calculates the standard error of the mean for a given data set x. It outputs a numeric value for the standard deviation of x 
over the root of the number of values in x. It does not include a rounding function, but does remove NA values.
}
\details{
As mentioned above, the Standard Error of the Mean is calculated from the standard deviation of x, and the square root of n, the 
number of measurements in x. The formula used is as follows:
         SEM= sd(x)/sqrt(n)
}
\examples{
Calculating the Standard Error of the Mean for columnName in NameOfDataFile data set
         file <-read.csv("NameOfDataFile.csv")
         SEM(file$columName)
         #value of SEM (numeric)
         
         Calculating the Standard Error of the Mean for speed in the built-in data set 'cars'.
         SEM(cars$speed)
         0.7477858 #Standard error of the mean of speed data in the built-in data set cars
}
