% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{Summary Statistics Function}
\usage{
sum_stats(x)
}
\arguments{
\item{x}{the data set for which the summary statistics is desired}
}
\description{
This function calculates the mean, median, range, variance, and standard deviation, and prints them as an appropriately labelled parcel of 
text. It also adheres to the rounding conventions (to one greater decimal than measured) established in class. Additionally, NA values are 
removed within the function.
}
\examples{
x<- c(0.1,0.2,0.3,0.4) #data is measured to one decimal place
         sum_stats(x)
Summary Statistics:
mean = 0.25 #results are given to two decimal places (decimals measured to + 1)
median = 0.25
range = 0.3 #R does not typically include terminal 0s in decimals reported from calculations
variance = 0.02
standard deviation = 0.13
}
