% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempStats.R
\name{TempStats}
\alias{TempStats}
\title{Pescadero Temperature Data Analysis Function}
\usage{
TempStats(x,t,d)
}
\arguments{
\item{x}{file name of the csv containing the desired temperature data (eg: "ISS_001_log.csv")}

\item{t}{the timestamp of the desired start time, including the year, month, day, hour, minute, and second. (eg. 2017-10-31T19:00:12.656Z)}

\item{d}{the duration, in minutes, of the data one wishes to analyze}
}
\description{
This function depends on the package 'lubridate', as well as the functions sum_stats() and SEM().
Lubridate is not included in this package, PescStats, but sum_stats() and SEM() are. This function 
prints the mean, median, range, variance, standard deviation, and standard error of the mean for data 
in the "temperature" column of a .csv file for a given time range, starting at a given time. It also 
generates a bar plot of the temperatures for that range.
}
\examples{
         TempStats("ISS_001_log.csv","2017-10-31 14:48:00",10)
Temperature Data from 2017-10-31 14:48:00 to 2017-10-31 14:58:00
Summary Statistics:
mean = 40.627
median = 40.48
range = 16.78
variance = 12.122
standard deviation = 3.482
Standard Error of the Mean = 0.176
 
         TempStats("temp_log_002.csv","2017-10-31T20:53:26.428Z",15)
Temperature Data from 2017-10-31 20:53:26 to 2017-10-31 21:08:26
Summary Statistics:
 mean = 9.511
 median = 9.27
 range = 10.4
 variance = 10.987
 standard deviation = 3.315
 Standard Error of the Mean = 0.135
}
